+++ oauth2/src/main/scala/org/broadinstitute/dsde/workbench/oauth2/OpenIDConnectAkkaHttpOps.scala
@@ -2,12 +2,17 @@ package org.broadinstitute.dsde.workbench.oauth2
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
+import akka.http.scaladsl.marshalling.{Marshaller, ToEntityMarshaller}
+import akka.http.scaladsl.model.ContentTypes._
 import akka.http.scaladsl.model.HttpMethods.POST
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import akka.stream.scaladsl.Flow
 import akka.util.ByteString
+import io.circe.Encoder
+import io.circe.syntax._
+import org.broadinstitute.dsde.workbench.oauth2.OpenIDConnectAkkaHttpOps.ConfigurationResponse
 
 import java.nio.file.Paths
 
@@ -20,7 +25,7 @@ class OpenIDConnectAkkaHttpOps(private val config: OpenIDConnectConfiguration) {
         get {
           parameterSeq { params =>
             val newQuery = Uri.Query(config.processAuthorizeQueryParams(params): _*)
-            val newUri = Uri(config.getAuthorizationEndpoint).withQuery(newQuery)
+            val newUri = Uri(config.providerMetadata.authorizeEndpoint).withQuery(newQuery)
             redirect(newUri, StatusCodes.Found)
           }
         }
@@ -31,11 +36,18 @@ class OpenIDConnectAkkaHttpOps(private val config: OpenIDConnectConfiguration) {
               complete {
                 val newRequest = HttpRequest(
                   POST,
-                  uri = Uri(config.getTokenEndpoint),
+                  uri = Uri(config.providerMetadata.tokenEndpoint),
                   entity = FormData(config.processTokenFormFields(fields): _*).toEntity
                 )
                 Http().singleRequest(newRequest)
               }
+            }
+          }
+        } ~
+        path("configuration") {
+          get {
+            complete {
+              ConfigurationResponse(config.authorityEndpoint, config.clientId)
             }
           }
         }
@@ -66,4 +78,18 @@ class OpenIDConnectAkkaHttpOps(private val config: OpenIDConnectConfiguration) {
         }
       }
   }
+}
+
+object OpenIDConnectAkkaHttpOps {
+  case class ConfigurationResponse(authorityEndpoint: String, clientId: ClientId)
+
+  implicit final def marshaller[A: Encoder]: ToEntityMarshaller[A] =
+    Marshaller.withFixedContentType(`application/json`) { a =>
+      HttpEntity(`application/json`, a.asJson.noSpaces)
+    }
+
+  implicit val clientIdEncoder: Encoder[ClientId] =
+    Encoder.encodeString.contramap(_.value)
+  implicit val configurationResponseEncoder: Encoder[ConfigurationResponse] =
+    Encoder.forProduct2("authorityEndpoint", "clientId")(x => (x.authorityEndpoint, x.clientId))
 }