+++ oauth2/src/test/scala/org/broadinstitute/dsde/workbench/oauth2/mock/FakeOpenIDConnectConfiguration.scala
@@ -1,13 +1,16 @@
 package org.broadinstitute.dsde.workbench.oauth2.mock
 
-import org.broadinstitute.dsde.workbench.oauth2.OpenIDConnectConfiguration
+import org.broadinstitute.dsde.workbench.oauth2.{ClientId, OpenIDConnectConfiguration, OpenIDProviderMetadata}
 
 class FakeOpenIDConnectConfiguration extends OpenIDConnectConfiguration {
-  override def getAuthorizationEndpoint: String = "some-auth-provider"
+  override def clientId: ClientId = ClientId("some-client")
+
+  override def authorityEndpoint: String = "https://fake"
+
+  override def providerMetadata: OpenIDProviderMetadata =
+    OpenIDProviderMetadata("fake-issuer", "fake-authorize", "fake-token")
 
   override def processAuthorizeQueryParams(params: Seq[(String, String)]): Seq[(String, String)] = params
-
-  override def getTokenEndpoint: String = "some-token-endpoint"
 
   override def processTokenFormFields(fields: Seq[(String, String)]): Seq[(String, String)] = fields